[v8-sandbox] Don't alert on the JS corruption API failing
We create alerts when the logs contain ": Assertion Failed" to catch
Swift assertions. We have cases where the ExplorationMutator runs into
issues when running the instrumented code that contains both the
instrumentation for the corrupt() call to cause in-sandbox corruption as
well as the exploration mutation running and then causing one of these
assert calls in JS to fail.

I don't think it should be considered a critical unexpected failure when
this is triggered.

Change-Id: I149132d9f081df62554d2e3c543fb63f9c299b71
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8719276
Auto-Submit: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Michael Achenbach <machenbach@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>