Add support for rest parameters in functions
This change adds a possibility of generation rest parameters.
The logic in 'randomParameters' was updated to generate a rest parameter with 20% chance whenever possible
The chance is an optional function argument, so we can modify that value in tests (and in the code, if necessary).

Bug: 458042811, 456162872
Change-Id: Idaf6446cfd248878134e7a449746260e44cfb74b
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8741997
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Pawel Krawczyk <pawkra@google.com>
Commit-Queue: Dominik Klemba <tacet@google.com>
Auto-Submit: Dominik Klemba <tacet@google.com>