Abort runtime-assisted mutations on FuzzIL translation errors
Previously, if we caught an error during translation of a
RuntimeAssistedMutator's Actions back to FuzzIL instructions, we simply
logged an error but otherwise continued the mutation. This, however, can
lead to invalid programs as we're essentially deleting the current
instruction (which may e.g. have produced an output needed later on).
This might then lead to various other crashes later on. Instead, we
should simply abort the mutation when we see such an error.

Bug: 468928010
Change-Id: Iefd1010d9c7bd72444d5be1258f81b3063f7b39b
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8849276
Reviewed-by: Michael Achenbach <machenbach@google.com>
Commit-Queue: Samuel Groß <saelo@google.com>